WIP: Add `gulp build`
The idea is to have labels for builds via a `--label` flag (or env var),
and to avoid changing existing stuff, `built/local` is still where
everything happens.  So there is always an "active" label which lives
there, and when you switch a label, `built/local` for the old label is
moved away and the one for the new label is renamed to it.

There are a few magical labels: `local` doesn't keep a build info
file (to keep it the same as `gulp local`); an empty-string label means
continue building the currently active label; and `git-branch` makes it
use the current branch name as the label.

The build info that is kept for each build is the label name, and a hash
of all sources (`src/**/*`).  If a `build` is done and the sources are
the same, then the actual build is skipped -- using hashes is needed to
make it easy to switch branches, since git touches modified files on
every branch switch.  (This is not done for `local` since it doesn't get
an info file.)